Added base players.games._GameWeapons class to be used when the game is not supported.
Added DialogType to messages the package.  create_message wrapper classes will soon be added as well.

Moved auth.base functionality to auth.providers as the exposed class in that module is specifically designed to be inherited by all included providers.

Changed auth.commands._auth_commands values to be methods of the instance instead of functions.