Added documentation for CallingConvention
Added the macro CHECK_OVERRIDE to check if a method was overriden. This
is because we had to get rid of Boost's pure_virtual function, because
that generates two functions, which fucks up the documentation generated
with Sphinx. Moreover, we can now raise a NotImplementedError, which is
better than a RuntimeError.